backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
langchain==0.0.350
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
python-dotenv==1.0.0
pydantic==2.5.1
python-multipart==0.0.6
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4